chore: create re-usable fuel selection component
