+++ ggircs-app/app/schema/schema.json
@@ -325,6 +325,18 @@
               "deprecationReason": null
             },
             {
+              "name": "ctaRateUnits",
+              "description": "The units of measure for the fuel type as defined in the carbon tax act https://www.bclaws.gov.bc.ca/civix/document/id/complete/statreg/08040_01",
+              "args": [],
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "isDeprecated": false,
+              "deprecationReason": null
+            },
+            {
               "name": "fuelCarbonTaxDetailsByCarbonTaxActFuelTypeId",
               "description": "Reads and enables pagination through a set of `FuelCarbonTaxDetail`.",
               "args": [
@@ -551,6 +563,18 @@
               "deprecationReason": null
             },
             {
+              "name": "metadata",
+              "description": "Column contains metadata pertaining to each fuel type",
+              "args": [],
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "isDeprecated": false,
+              "deprecationReason": null
+            },
+            {
               "name": "rowId",
               "description": "The internal primary key",
               "args": [],
@@ -587,6 +611,26 @@
             {
               "name": "carbonTaxFuelType",
               "description": "Checks for equality with the object’s `carbonTaxFuelType` field.",
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "defaultValue": null
+            },
+            {
+              "name": "ctaRateUnits",
+              "description": "Checks for equality with the object’s `ctaRateUnits` field.",
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "defaultValue": null
+            },
+            {
+              "name": "metadata",
+              "description": "Checks for equality with the object’s `metadata` field.",
               "type": {
                 "kind": "SCALAR",
                 "name": "String",
@@ -644,6 +688,16 @@
               "defaultValue": null
             },
             {
+              "name": "ctaRateUnits",
+              "description": "Filter by the object’s `ctaRateUnits` field.",
+              "type": {
+                "kind": "INPUT_OBJECT",
+                "name": "StringFilter",
+                "ofType": null
+              },
+              "defaultValue": null
+            },
+            {
               "name": "fuelCarbonTaxDetailsByCarbonTaxActFuelTypeId",
               "description": "Filter by the object’s `fuelCarbonTaxDetailsByCarbonTaxActFuelTypeId` relation.",
               "type": {
@@ -679,6 +733,16 @@
               "type": {
                 "kind": "SCALAR",
                 "name": "Boolean",
+                "ofType": null
+              },
+              "defaultValue": null
+            },
+            {
+              "name": "metadata",
+              "description": "Filter by the object’s `metadata` field.",
+              "type": {
+                "kind": "INPUT_OBJECT",
+                "name": "StringFilter",
                 "ofType": null
               },
               "defaultValue": null
@@ -745,6 +809,26 @@
                 }
               },
               "defaultValue": null
+            },
+            {
+              "name": "ctaRateUnits",
+              "description": "The units of measure for the fuel type as defined in the carbon tax act https://www.bclaws.gov.bc.ca/civix/document/id/complete/statreg/08040_01",
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "defaultValue": null
+            },
+            {
+              "name": "metadata",
+              "description": "Column contains metadata pertaining to each fuel type",
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "defaultValue": null
             }
           ],
           "interfaces": null,
@@ -760,6 +844,26 @@
             {
               "name": "carbonTaxFuelType",
               "description": "Fuel type defined by the carbon tax act",
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "defaultValue": null
+            },
+            {
+              "name": "ctaRateUnits",
+              "description": "The units of measure for the fuel type as defined in the carbon tax act https://www.bclaws.gov.bc.ca/civix/document/id/complete/statreg/08040_01",
+              "type": {
+                "kind": "SCALAR",
+                "name": "String",
+                "ofType": null
+              },
+              "defaultValue": null
+            },
+            {
+              "name": "metadata",
+              "description": "Column contains metadata pertaining to each fuel type",
               "type": {
                 "kind": "SCALAR",
                 "name": "String",
@@ -997,6 +1101,18 @@
               "deprecationReason": null
             },
             {
+              "name": "CTA_RATE_UNITS_ASC",
+              "description": null,
+              "isDeprecated": false,
+              "deprecationReason": null
+            },
+            {
+              "name": "CTA_RATE_UNITS_DESC",
+              "description": null,
+              "isDeprecated": false,
+              "deprecationReason": null
+            },
+            {
               "name": "ID_ASC",
               "description": null,
               "isDeprecated": false,
@@ -1004,6 +1120,18 @@
             },
             {
               "name": "ID_DESC",
+              "description": null,
+              "isDeprecated": false,
+              "deprecationReason": null
+            },
+            {
+              "name": "METADATA_ASC",
+              "description": null,
+              "isDeprecated": false,
+              "deprecationReason": null
+            },
+            {
+              "name": "METADATA_DESC",
               "description": null,
               "isDeprecated": false,
               "deprecationReason": null
